refactor: application port 패키지 구조 통합
